ADD_UNIT_TESTS_FOR_TRUCKS
This commit contains unit tests for trucks. Changes (+ explanations) still have to be made, but all tests are working and a general coverage of 64.62 is reached if you count all the files associated with each other. If we only count repo, controller, schema, and service we have a coverage of 82.71%