4_B_UPDATE_DELIVERY_FRONTEND_TO_ALLOW_PAGING
This commit updates the frontend of the deliveries to include paging buttons and a dropdown menu to select the amount of items you want per page.
It also has logic that performs certain calculations to (for example) get the amount of pages that we can have when selecting a certain amount of items. It does this by retrieving all the delivery data and get the length of this.

Additionally, the service is also updated to call the new backend route (/paged).