4_B_FIX_DELIVERY_PAGING_BUG
This commit fixes the bug that messed up the paging buttons when a user is on a page different than 2 and changes the amount of items they want to see on each page. The previous button value was not updated which lead to wrong values and bugs occuring on the page.