3.3.1.1.a_UPDATE_USER_BACKEND
This commit updates the user logic in the backend.
A new encryption is being used for the password so that Auth0 (task 3.3.1.1.c) can verify the credentials (as they used another encryption method).
Logging in by the api is NOT possible anymore.
This commit also added logic to retrieve personal user data based on email (except the password). This data is not protected yet.