3.3.1.1.c_IMPLEMENT_SSO
This is a big commit that changes the whole frontend.

This commit includes the Auth0 sample project with all the files from the previously build frontend.

Those file are also modified as now there is an additional check to see if the user is authenticated, and a role check to hide certain pages from certain users.

Only missing feature is the addition of users.